refactor: 에러 페이지 및 클라이언트 레이아웃 업데이트
- ErrorPage 컴포넌트에 스타일 및 버튼 추가
- ClientLayout에 ErrorBoundary 및 QueryErrorResetBoundary 통합
