+++ package.json
@@ -26,6 +26,7 @@
     "next": "15.1.7",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-error-boundary": "^6.0.0",
     "react-hook-form": "^7.55.0",
     "react-loading-skeleton": "^3.5.0",
     "react-simplikit": "^0.0.30",
