+++ src/app/_layouts/header/useHeader.ts
@@ -1,7 +1,7 @@
 import { signOutUser } from '@/features/auth/apis';
 import { useUser } from '@/features/user/hooks';
 
-export default function useHeader() {
+export function useHeader() {
   const { user, setUser } = useUser();
   const userEmail = user?.email ?? '';
