--- src/shared/components/layouts/Portal.tsx
@@ -1,11 +1,0 @@
-import { createPortal } from 'react-dom';
-
-export default function Portal({ children }: { children: React.ReactNode }) {
-  const modalRoot = document.getElementById('modal-root');
-
-  if (!modalRoot) {
-    return null;
-  }
-
-  return createPortal(children, modalRoot);
-}
