+++ src/shared/components/layouts/index.ts
@@ -1,4 +1,3 @@
-export { default as Header } from './header/Header';
 export { default as CardLayout } from './CardLayout';
 export { default as Modal } from './Modal';
 export { default as CheckoutSteps } from './CheckoutSteps';
